Refactor RhinoRenderTool into partial classes
Split the 2,020-line RhinoRenderTool.cs into 5 files for better
maintainability and code navigation:

- RhinoRenderTool.cs: Core dispatch, ToolInfo, display/camera/zoom/modes/render
- RhinoRenderTool.Settings.cs: Background, ground plane, sun, and skylight settings
- RhinoRenderTool.Views.cs: Named view actions (save, load, list, delete)
- RhinoRenderTool.Lights.cs: Light management (add, list, set, delete)
- RhinoRenderTool.Materials.cs: Material and environment actions

No functional changes - the MCP tool API remains identical.
